+++ src/workflows/airqo_etl_utils/airqo_utils.py
@@ -729,64 +729,6 @@ class AirQoDataUtils:
         )
 
     @staticmethod
-    def get_devices_kafka(group_id: str) -> pd.DataFrame:
-        """
-        Fetches and returns a DataFrame of devices from the 'devices-topic' Kafka topic.
-
-        Args:
-            group_id (str): The consumer group ID used to track message consumption from the topic.
-
-        Returns:
-            pd.DataFrame: A DataFrame containing the list of devices, where each device is represented as a row.
-                      If any errors occur during the process, an empty DataFrame is returned.
-        """
-        from airqo_etl_utils.message_broker_utils import MessageBrokerUtils
-        from confluent_kafka import KafkaException
-        import json
-
-        broker = MessageBrokerUtils()
-        devices_list: List = []
-
-        for message in broker.consume_from_topic(
-            topic="devices-topic",
-            group_id=group_id,
-            auto_offset_reset="earliest",
-            auto_commit=False,
-        ):
-            try:
-                key = message.get("key", None)
-                try:
-                    value = json.loads(message.get("value", None))
-                except json.JSONDecodeError as e:
-                    logger.exception(f"Error decoding JSON: {e}")
-                    continue
-
-                if not key or not value.get("device_id"):
-                    logger.warning(
-                        f"Skipping message with key: {key}, missing 'device_id'."
-                    )
-                    continue
-
-                devices_list.append(value)
-            except KafkaException as e:
-                logger.exception(f"Error while consuming message: {e}")
-            continue
-
-        try:
-            devices = pd.DataFrame(devices_list)
-        except Exception as e:
-            logger.exception(f"Failed to convert consumed messages to DataFrame: {e}")
-            # Return empty DataFrame on failure
-            devices = pd.DataFrame()
-
-        if "device_name" in devices.columns.tolist():
-            devices.drop_duplicates(subset=["device_name"], keep="last")
-        elif "device_id" in devices.columns.tolist():
-            devices.drop_duplicates(subset=["device_id"], keep="last")
-
-        return devices
-
-    @staticmethod
     def extract_devices_with_uncalibrated_data(
         start_date, table: str = None, network: DeviceNetwork = DeviceNetwork.AIRQO
     ) -> pd.DataFrame:
